moved templates folder
